requirements.txt
# 核心依赖
requests>=2.32.5,<3.0.0
pytz>=2025.2,<2026.0
PyYAML>=6.0.3,<7.0.0

# MCP 服务器
fastmcp>=2.12.0,<2.14.0
websockets>=13.0,<14.0

# RSS 订阅源支持
feedparser>=6.0.10,<7.0.0

# API 服务器（新增）
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
pydantic>=2.5.0,<3.0.0

# 大模型 API 客户端（新增）
openai>=1.10.0,<2.0.0

# 测试框架
pytest>=7.4.0,<9.0.0
pytest-mock>=3.11.1,<4.0.0
